[Edit] UserRepositoryTest 변수, @BeforeEach 추상화
