+++ src/test/java/com/inq/wishhair/wesharewishhair/common/base/RepositoryTest.java
@@ -1,7 +1,30 @@
 package com.inq.wishhair.wesharewishhair.common.base;
 
+import com.inq.wishhair.wesharewishhair.domain.user.User;
+import com.inq.wishhair.wesharewishhair.domain.user.repository.UserRepository;
+import com.inq.wishhair.wesharewishhair.fixture.UserFixture;
+import org.junit.jupiter.api.BeforeEach;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 
 @DataJpaTest
 public abstract class RepositoryTest {
+
+    @Autowired
+    protected UserRepository userRepository;
+
+    protected User userA;
+    protected User userB;
+    protected User userC;
+
+    @BeforeEach
+    void saveUsers() {
+        userA = UserFixture.A.toEntity();
+        userB = UserFixture.B.toEntity();
+        userC = UserFixture.C.toEntity();
+
+        userRepository.save(userA);
+        userRepository.save(userB);
+        userRepository.save(userC);
+    }
 }
