+++ src/test/java/com/inq/wishhair/wesharewishhair/user/repository/UserRepositoryTest.java
@@ -14,24 +14,6 @@ import static org.junit.jupiter.api.Assertions.*;
 @DisplayName("UserRepository Test")
 class UserRepositoryTest extends RepositoryTest {
 
-    @Autowired
-    private UserRepository userRepository;
-
-    private User userA;
-    private User userB;
-    private User userC;
-
-    @BeforeEach
-    void saveUsers() {
-        userA = UserFixture.A.toEntity();
-        userB = UserFixture.B.toEntity();
-        userC = UserFixture.C.toEntity();
-
-        userRepository.save(userA);
-        userRepository.save(userB);
-        userRepository.save(userC);
-    }
-
     @Test
     @DisplayName("회원 저장 테스트")
     void userSaveTest() {
